Move all AOP-related functionality to inject.aop package
