+++ src/main/php/inject/aop/MethodInvocation.class.php
@@ -1,4 +1,4 @@
-<?php namespace inject;
+<?php namespace inject\aop;
 
 use util\Objects;
