+++ src/main/php/inject/aop/ProxyProvider.class.php
@@ -1,18 +1,19 @@
-<?php namespace inject;
+<?php namespace inject\aop;
 
 use lang\XPClass;
+use inject\Injector;
 
 #[@generic(implements= ['var'])]
-class ProxyProvider extends \lang\Object implements Provider {
+class ProxyProvider extends \lang\Object implements \inject\Provider {
   protected $proxy;
   protected $injector;
 
   /**
    * Creates a new type provider
    *
-   * @param  lang.XPClass $type
+   * @param  var $type Either an XPClass instance or a string
    * @param  inject.Injector $injector
-   * @param  inject.MethodInterception $interception
+   * @param  inject.aop.MethodInterception $interception
    */
   public function __construct($type, Injector $injector, MethodInterception $interception) {
     $this->proxy= new Proxy(
