+++ src/test/php/inject/unittest/aop/InterceptionTest.class.php
@@ -1,20 +1,20 @@
-<?php namespace inject\unittest;
+<?php namespace inject\unittest\aop;
 
 use inject\Injector;
-use inject\ProxyProvider;
-use inject\MethodInvocation;
+use inject\aop\ProxyProvider;
+use inject\aop\MethodInvocation;
 use unittest\TestCase;
 
 class InterceptionTest extends TestCase {
 
   #[@test]
   public function _() {
-    $log= create('new util.collections.Vector<inject.MethodInvocation>');
+    $log= create('new util.collections.Vector<inject.aop.MethodInvocation>');
     $inject= new Injector();
     $inject->bind('inject.unittest.fixture.Storage', new ProxyProvider(
       'inject.unittest.fixture.FileSystem',
       $inject,
-      newinstance('inject.MethodInterception', [], [
+      newinstance('inject.aop.MethodInterception', [], [
         'invoke' => function($invocation) use($log) {
           $log[]= $invocation;
           return $invocation->proceed();
